chore: move tests to backend/
