+++ backend/tests/pre_start/test_db.py
@@ -1,0 +1,16 @@
+import logging
+
+import pytest
+
+from app.core.db import engine
+from app.utils.pre_start import init as init_db
+from app.utils.test_pre_start import init as connect_db
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+@pytest.mark.anyio
+async def test_init_db() -> None:
+    await connect_db(engine)
+    await init_db(engine)
